requirements_academic.txt
# Core Dependencies
torch>=1.13.0
torchvision>=0.14.0
timm==0.5.4
fvcore

# Data Processing and Augmentation
Pillow>=8.3.0
numpy>=1.21.0
opencv-python>=4.5.0

# Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Experiment Tracking and Logging
wandb>=0.12.0
tensorboard>=2.8.0
tqdm>=4.62.0

# Model Export and Deployment
coremltools>=5.0.0
onnx>=1.12.0
onnxruntime>=1.12.0

# Development and Testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Optional: for advanced features
# transformers>=4.20.0  # If using HuggingFace models
# accelerate>=0.12.0    # For distributed training

